refactor(web): split github-core/queries.ts into resource sub-modules behind a barrel
Decompose the 1593-line reads monolith into cohesive ./queries/* sub-modules
(user/org/repoRef/file/team/invitation/pages/releaseRun reads) with two leaf
modules: keys.ts (githubKeys + invalidateInviteQueries, imported by every read)
and shared.ts (sleep, fresh-repo retry, REPO_READ_CONCURRENCY, scoped log).
queries.ts is now a thin re-export barrel so all ~68 importers are unchanged.

Behavior-preserving: every export name preserved (tsc + full vitest green), the
data-layer no-cycle guard + boundaries + arch:validate stay green (leaf
discipline), and queries.test.ts still exercises the public surface via the
barrel. Largest sub-module is 316 lines (was 1593).
